rename QueryRunner to PanelQueryRunner
